examples/requirements.txt
nornir_pygnmi
nornir_utils